+++ login/index.html
@@ -35,7 +35,7 @@
                     </div>
                 </article>
 
-                <button type="button" class="login_button">로그인</button>
+                <button type="button" class="login_button" disabled>로그인</button>
             </div>
             <div class="simple_login">
                 <div class="simple_login_text">
@@ -62,5 +62,5 @@
     </section>
     
 </body>
-<script type="module" src="/Script/login_js.js"></script>
+<script type="module" src="/script/login.js"></script>
 </html>
