+++ signup/index.html
@@ -46,7 +46,7 @@
                         <img src="/images/TheEye.png" alt="Eye_for_password" ID="togglePassword_con" class="toggleEye">
                     </div>
                 </article>
-                <button type="button" class="login_button">회원가입</button>
+                <button type="button" class="login_button" disabled>회원가입</button>
             </div>
             <div class="simple_login">
                 <div class="simple_login_text">
@@ -71,7 +71,7 @@
             </div>
         </div>
     </section>
-    <script type="module" src="/Script/signup_js.js"></script>
+    <script type="module" src="/script/signup.js"></script>
 </body>
 
 </html>
